feat: specific position endpoint
