+++ packages/api/src/v1-queries/get-portfolio-positions/types.ts
@@ -4,6 +4,7 @@ export type V1PortfolioPosition = {
   id: string;
 
   type: 'LP' | 'Variable' | 'Fixed';
+  variant: 'matured' | 'settled' | 'active';
   creationTimestampInMS: number;
 
   ownerAddress: string;
@@ -14,24 +15,44 @@ export type V1PortfolioPosition = {
   fixLow: number;
   fixHigh: number;
 
-  tokenPriceUSD: number;
   notionalProvided: number;
   notionalTraded: number;
   notional: number;
   margin: number;
-
-  status: {
-    health: 'healthy' | 'danger' | 'warning';
-    variant: 'matured' | 'settled' | 'active';
-    currentFixed: number;
-    receiving: number;
-    paying: number;
-  };
 
   unrealizedPNL: number;
   realizedPNLFees: number;
   realizedPNLCashflow: number;
   realizedPNLTotal: number;
 
-  amm: V1Pool;
+  health: 'healthy' | 'danger' | 'warning';
+  receiving: number;
+  paying: number;
+
+  poolCurrentFixedRate: number;
+  pool: V1Pool;
+};
+
+export type V1HistoryTransaction = {
+  type:
+    | 'swap'
+    | 'mint'
+    | 'burn'
+    | 'margin-update'
+    | 'liquidation'
+    | 'settlement'
+    | 'maturity';
+  creationTimestampInMS: number;
+  notional: number;
+  paidFees: number;
+  fixedRate: number;
+  marginDelta: number;
+};
+
+export type V1PortfolioPositionDetails = V1PortfolioPosition & {
+  canEdit: boolean;
+  canSettle: boolean;
+  rolloverPoolId: null | string;
+
+  history: V1HistoryTransaction[];
 };
