--- packages/api/src/v1-queries/position-pnl/types.ts
@@ -1,7 +1,0 @@
-export type PositionPnL = {
-  realizedPnLFromSwaps: number;
-  realizedPnLFromFeesPaid: number;
-  realizedPnLFromFeesCollected: number;
-  unrealizedPnLFromSwaps: number;
-  fixedRateLocked: number;
-};
