+++ packages/api/src/v1v2-queries/get-portfolio-positions/extendV1Position.ts
@@ -1,11 +1,23 @@
-import { V1PortfolioPosition } from '../../v1-queries/get-portfolio-positions/types';
+import {
+  V1PortfolioPosition,
+  V1PortfolioPositionDetails,
+} from '../../v1-queries/get-portfolio-positions/types';
 import { extendV1Pool } from '../get-pools/extendV1Pool';
-import { V1V2PortfolioPosition } from './types';
+import { V1V2PortfolioPosition, V1V2PortfolioPositionDetails } from './types';
 
 export const extendV1Position = (
   p: V1PortfolioPosition,
 ): V1V2PortfolioPosition => ({
   ...p,
   accountId: '',
-  amm: extendV1Pool(p.amm),
+  pool: extendV1Pool(p.pool),
+});
+
+export const extendV1PositionDetails = (
+  p: V1PortfolioPositionDetails,
+): V1V2PortfolioPositionDetails => ({
+  ...p,
+  accountId: '',
+  pool: extendV1Pool(p.pool),
+  rolloverMaturityTimestamp: null,
 });
