+++ packages/api/src/v1v2-queries/get-portfolio-positions/extendV2Position.ts
@@ -1,10 +1,22 @@
-import { V2PortfolioPosition } from '../../v2-queries/get-portfolio-positions/types';
+import {
+  V2PortfolioPosition,
+  V2PortfolioPositionDetails,
+} from '../../v2-queries/get-portfolio-positions/types';
 import { extendV2Pool } from '../get-pools/extendV2Pool';
-import { V1V2PortfolioPosition } from './types';
+import { V1V2PortfolioPosition, V1V2PortfolioPositionDetails } from './types';
 
 export const extendV2Position = (
   p: V2PortfolioPosition,
 ): V1V2PortfolioPosition => ({
   ...p,
-  amm: extendV2Pool(p.amm),
+  pool: extendV2Pool(p.pool),
+});
+
+export const extendV2PositionDetails = (
+  p: V2PortfolioPositionDetails,
+): V1V2PortfolioPositionDetails => ({
+  ...p,
+  accountId: '',
+  pool: extendV2Pool(p.pool),
+  rolloverPoolId: null,
 });
