+++ packages/api/src/v2-queries/get-portfolio-positions/types.ts
@@ -3,10 +3,12 @@ import { V2Pool } from '../get-pools/types';
 export type V2PortfolioPosition = {
   id: string;
 
+  type: 'LP' | 'Variable' | 'Fixed';
+  variant: 'matured' | 'settled' | 'active';
+  creationTimestampInMS: number;
+
   accountId: string;
   ownerAddress: string;
-  type: 'LP' | 'Variable' | 'Fixed';
-  creationTimestampInMS: number;
 
   tickLower: number;
   tickUpper: number;
@@ -14,24 +16,44 @@ export type V2PortfolioPosition = {
   fixLow: number;
   fixHigh: number;
 
-  tokenPriceUSD: number;
   notionalProvided: number;
   notionalTraded: number;
   notional: number;
   margin: number;
-
-  status: {
-    health: 'healthy' | 'danger' | 'warning';
-    variant: 'matured' | 'settled' | 'active';
-    currentFixed: number;
-    receiving: number;
-    paying: number;
-  };
 
   unrealizedPNL: number;
   realizedPNLFees: number;
   realizedPNLCashflow: number;
   realizedPNLTotal: number;
 
-  amm: V2Pool;
+  health: 'healthy' | 'danger' | 'warning';
+  receiving: number;
+  paying: number;
+
+  poolCurrentFixedRate: number;
+  pool: V2Pool;
+};
+
+export type V2HistoryTransaction = {
+  type:
+    | 'swap'
+    | 'mint'
+    | 'burn'
+    | 'margin-update'
+    | 'liquidation'
+    | 'settlement'
+    | 'maturity';
+  creationTimestampInMS: number;
+  notional: number;
+  paidFees: number;
+  fixedRate: number;
+  marginDelta: number;
+};
+
+export type V2PortfolioPositionDetails = V2PortfolioPosition & {
+  canEdit: boolean;
+  canSettle: boolean;
+  rolloverMaturityTimestamp: null | number;
+
+  history: V2HistoryTransaction[];
 };
