+++ packages/commons/src/v1-ids/encodeV1PoolId.ts
@@ -1,0 +1,10 @@
+import { V1PoolIdData } from './types';
+
+export const encodeV1PoolId = ({
+  chainId,
+  vammAddress,
+}: V1PoolIdData): string => {
+  const poolId = `${chainId}_${vammAddress.toLowerCase()}_v1`;
+
+  return poolId;
+};
