+++ packages/commons/src/v1-ids/types.ts
@@ -1,0 +1,14 @@
+import { SupportedChainId } from '../provider';
+
+export type V1PoolIdData = {
+  chainId: SupportedChainId;
+  vammAddress: string;
+};
+
+export type V1PositionIdData = {
+  chainId: SupportedChainId;
+  vammAddress: string;
+  ownerAddress: string;
+  tickLower: number;
+  tickUpper: number;
+};
