+++ packages/v2-bigquery/src/positions-table/index.ts
@@ -3,4 +3,5 @@ export * from './operations/updateEntry';
 export * from './operations/pullEntry';
 export * from './operations/pullAccountPositionEntries';
 export * from './operations/pullLpPositionEntries';
+export * from './operations/pullEntryById';
 export { PositionEntry } from './specific';
